move graph to verifier
